refactor to separate pages and components
